+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Legenda for the issue kind:
 - [#563](https://github.com/ariadne-cps/ariadne/issues/563) (C) Rename the 'output' module into the 'io' module to support future classes related to input
 - [#539](https://github.com/ariadne-cps/ariadne/issues/539) (F) A segmentation fault sometimes would be issued when terminating the executable, due to logging
 - [#557](https://github.com/ariadne-cps/ariadne/issues/557) (F) Fix behavior of StopWatch utility for concurrent code, enhance the class for choosing a duration type
+- [#567](https://github.com/ariadne-cps/ariadne/issues/567) (F) Fix problem with state_time space creation when state space already contains the 't' variable
 
 ### 2.2